# openapi/openapi.yaml
...
openapi: "3.0.1"
info:
  title: "Location Service API"
  description: "Location service provides a REST API mainly for purposes of searching\
    \ localities across Slovak Republic based on GPS coordinates, with possibility\
    \ of extending searched area by other countries in the future."
  contact:
    url: "https://sk.linkedin.com/in/martin-lokaj-94a679154"
    email: "mrtn.lkj@gmail.com"
  license:
    name: "Licensed by Martin Lokaj"
  version: "0.0.1-SNAPSHOT"
servers:
- url: "http://localhost"
  description: "Generated server url"
...
paths:
  /api/v1/update:
    get:
      tags:
      - "UpdateRecordController"
      summary: "1101 - Get update records"
      description: "Returns list of history update records. Possibility to filter\
        \ results by specified query parameters."
      operationId: "getUpdateRecords"
      parameters:
      - name: "status"
        in: "query"
        required: false
        schema:
          type: "string"
          enum:
          - "RUNNING"
          - "FINISHED"
          - "FAILED"
        example: "FINISHED"
      - name: "trigger"
        in: "query"
        required: false
        schema:
          type: "string"
          enum:
          - "SCHEDULED_UPDATE"
          - "SCHEDULED_STARTUP_UPDATE"
          - "SCHEDULED_RETRY_UPDATE"
          - "MANUAL_UPDATE"
        example: "MANUAL_UPDATE"
      - name: "url"
        in: "query"
        required: false
        schema:
          type: "string"
        example: "https://download.geofabrik.de/europe/slovakia-latest.osm.pbf"
      - name: "dateStartedFrom"
        in: "query"
        required: false
        schema:
          type: "string"
          format: "date-time"
        example: "2022-05-01T00:00:00Z"
      - name: "limit"
        in: "query"
        description: "Limit returned records by specifying number of items to skip\
          \ before returning the results."
        required: false
        schema:
          minimum: 1
          type: "integer"
          format: "int64"
        example: 10
      - name: "offset"
        in: "query"
        description: "Offset returned records by specified number of records to skip\
          \ before returning the results."
        required: false
        schema:
          minimum: 0
          type: "integer"
          format: "int64"
        example: 0
      responses:
        "400":
          description: "Bad request.\n List of possible error codes: <ul><li>LS0001\
            \ - Request is not valid</li></ul>"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                LS0001:
                  description: "LS0001"
                  value:
                    errorMessage:
                      referenceId: "70b2e0fc-c656-4d6d-9e3e-0a85f43c3b91"
                      message: "Constraint violation (cause: org.springframework.validation.BeanPropertyBindingResult:\
                        \ 1 errors\nField error in object 'updatesFilter' on field\
                        \ 'limit': rejected value [-5]; codes [Min.updatesFilter.limit,Min.limit,Min.java.lang.Long,Min];\
                        \ arguments [org.springframework.context.support.DefaultMessageSourceResolvable:\
                        \ codes [updatesFilter.limit,limit]; arguments []; default\
                        \ message [limit],1]; default message [must be greater than\
                        \ or equal to 1])"
                    businessExceptions:
                    - code: "LS0001"
                      message: "Constraint violation"
                    fieldValidations:
                    - field: "limit"
                      message: "must be greater than or equal to 1"
        "200":
          description: "Success."
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GetUpdateRecordsResponse"
              examples:
                GetUpdateRecordsResponse:
                  description: "Updates found by search criteria"
                  value:
                    records:
                    - updateId: 11256
                      startedTime: "2022-04-05T19:10:17.155325Z"
                      finishedTime: "2022-04-05T19:51:40.327389Z"
                      dataDownloadUrl: "https://download.geofabrik.de/europe/slovakia-latest.osm.pbf"
                      status: "FINISHED"
                      trigger: "SCHEDULED_STARTUP_UPDATE"
                      description: null
                      failedReason: null
                    - updateId: 11255
                      startedTime: "2022-04-05T19:09:34.355455Z"
                      finishedTime: "2022-04-05T19:54:59.055674Z"
                      dataDownloadUrl: "https://download.geofabrik.de/europe/slovakia-latest.osm.pbf"
                      status: "FAILED"
                      trigger: "SCHEDULED_STARTUP_UPDATE"
                      description: null
                      failedReason: "Force kill stuck update - running more than 120\
                        \ minutes."
                    recordsCount: 2
                GetUpdateRecordsResponse - not found:
                  description: "No update records by search criteria was found"
                  value:
                    records: []
                    recordsCount: 0
        "500":
          description: "Internal server error.\n List of possible error codes:<ul><li>LS0004\
            \ - PostgreSQL JDBC connection failure</li></ul>"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                LS0004:
                  description: "LS0004"
                  value:
                    errorMessage:
                      referenceId: "efee2335-7e32-4be8-bd3b-65f2d06b6970"
                      message: "PostgreSQL JDBC connection failure. (cause: Failed\
                        \ to obtain JDBC Connection; nested exception is org.postgresql.util.PSQLException:\
                        \ FATAL: database \"xyz\" does not exist)"
                    businessExceptions:
                    - code: "LS0004"
                      message: "PostgreSQL JDBC connection failure."
                    fieldValidations: []
    post:
      tags:
      - "UpdateRecordController"
      summary: "1104 -  Endpoint that triggers manual update"
      description: "Performs a manual update of location data. This operation arranges\
        \ download of the OSM data file from specified server, import raw OSM data\
        \ to DB with osm2pgsql and prepare new version of location data."
      operationId: "manualUpdate"
      requestBody:
        description: "Optional request body that defines additional info to update."
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualUpdateRequest"
            examples:
              Manual update request with default settings.:
                description: "Manual update request with default settings."
                value:
                  url: null
                  description: "Trying to run manual update of locations data. Default\
                    \ data updater setting will be applied."
              Manual update request with specified custom url.:
                description: "Manual update request with specified custom url."
                value:
                  url: "https://download.geofabrik.de/europe/slovakia-latest.osm.pbf"
                  description: "Trying to run manual data update of locations data\
                    \ with specified custom URL for OSM data file download."
      responses:
        "200":
          description: "Success."
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UpdateRecordResponse"
              examples:
                Update record, response of processing manual update.:
                  description: "Update record, response of processing manual update."
                  value:
                    updateId: 1
                    startedTime: "2022-01-01T12:00:00.000Z"
                    finishedTime: null
                    dataDownloadUrl: "https://download.geofabrik.de/europe/slovakia-latest.osm.pbf"
                    status: "RUNNING"
                    trigger: "MANUAL_UPDATE"
                    description: "Trying to run manual update of locations data. Default\
                      \ data updater setting will be applied."
        "500":
          description: "Internal server error.\n List of possible error codes:<ul><li>LS0004\
            \ - PostgreSQL JDBC connection failure</li></ul>"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                LS0004:
                  description: "LS0004"
                  value:
                    errorMessage:
                      referenceId: "efee2335-7e32-4be8-bd3b-65f2d06b6970"
                      message: "PostgreSQL JDBC connection failure. (cause: Failed\
                        \ to obtain JDBC Connection; nested exception is org.postgresql.util.PSQLException:\
                        \ FATAL: database \"xyz\" does not exist)"
                    businessExceptions:
                    - code: "LS0004"
                      message: "PostgreSQL JDBC connection failure."
                    fieldValidations: []
                LS0101:
                  description: "LS0101"
                  value:
                    errorMessage:
                      referenceId: "10b0c42b-58db-47e6-8f34-742881779ed9"
                      message: "Update is already in progress."
                    businessExceptions:
                    - code: "LS0101"
                      message: "Update is already in progress."
                    fieldValidations: []
        "400":
          description: "Bad request.\n List of possible error codes: <ul><li>LS0001\
            \ - Request is not valid</li></ul>"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                LS0001:
                  description: "LS0001"
                  value:
                    errorMessage:
                      referenceId: "fc3510e1-f8b1-4131-ba78-b02178e83b40"
                      message: "Constraint violation (cause: Validation failed for\
                        \ argument [0] in public org.springframework.http.ResponseEntity<?>\
                        \ sk.uniza.locationservice.controller.UpdateController.manualUpdate(sk.uniza.locationservice.controller.bean.request.ManualUpdateRequest):\
                        \ [Field error in object 'manualUpdateRequest' on field 'url':\
                        \ rejected value [httpsinvalid]; codes [URL.manualUpdateRequest.url,URL.url,URL.java.lang.String,URL];\
                        \ arguments [org.springframework.context.support.DefaultMessageSourceResolvable:\
                        \ codes [manualUpdateRequest.url,url]; arguments []; default\
                        \ message [url],[Ljavax.validation.constraints.Pattern$Flag;@602a3c41,,-1,,.*];\
                        \ default message [must be a valid URL]] )"
                    businessExceptions:
                    - code: "LS0001"
                      message: "Constraint violation"
                    fieldValidations:
                    - field: "url"
                      message: "must be a valid URL"
  /api/v1/update/{updateId}:
    get:
      tags:
      - "UpdateRecordController"
      summary: "1102 - Get update record by id"
      description: "Returns a single update record for specified unique identifier\
        \ - updateId, response extended by update step details."
      operationId: "getUpdateRecordById"
      parameters:
      - name: "updateId"
        in: "path"
        description: "Unique updateId identifier."
        required: true
        schema:
          type: "integer"
          format: "int64"
        example: "11256"
      responses:
        "200":
          description: "Success."
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UpdateRecordResponse"
              examples:
                UpdateRecordResponse:
                  description: "Update by specified id"
                  value:
                    updateId: 11256
                    startedTime: "2022-04-05T19:10:17.155325Z"
                    finishedTime: "2022-04-05T19:51:40.327389Z"
                    dataDownloadUrl: "https://download.geofabrik.de/europe/slovakia-latest.osm.pbf"
                    status: "FINISHED"
                    trigger: "SCHEDULED_STARTUP_UPDATE"
                    description: null
                    failedReason: null
                UpdateRecordResponse - not found:
                  description: "No update found for specified id"
        "500":
          description: "Internal server error.\n List of possible error codes:<ul><li>LS0004\
            \ - PostgreSQL JDBC connection failure</li></ul>"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                LS0004:
                  description: "LS0004"
                  value:
                    errorMessage:
                      referenceId: "efee2335-7e32-4be8-bd3b-65f2d06b6970"
                      message: "PostgreSQL JDBC connection failure. (cause: Failed\
                        \ to obtain JDBC Connection; nested exception is org.postgresql.util.PSQLException:\
                        \ FATAL: database \"xyz\" does not exist)"
                    businessExceptions:
                    - code: "LS0004"
                      message: "PostgreSQL JDBC connection failure."
                    fieldValidations: []
        "400":
          description: "Bad request.\n List of possible error codes: <ul><li>LS0001\
            \ - Request is not valid</li></ul>"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                LS0001:
                  description: "LS0001"
                  value:
                    errorMessage:
                      referenceId: "5dcc4e1e-1ba6-4d83-a429-c1a44711594a"
                      message: "Constraint violation (cause: Failed to convert value\
                        \ of type 'java.lang.String' to required type 'java.lang.Long';\
                        \ nested exception is java.lang.NumberFormatException: For\
                        \ input string: \"sfasf\")"
                    businessExceptions:
                    - code: "LS0001"
                      message: "Constraint violation"
                    fieldValidations:
                    - field: "locationId"
                      message: "Failed to convert value of type 'java.lang.String'\
                        \ to required type 'java.lang.Long'; nested exception is java.lang.NumberFormatException:\
                        \ For input string: \"sfasf\""
  /api/v1/update/latest:
    get:
      tags:
      - "UpdateRecordController"
      summary: "1103 - Get latest update"
      description: "Returns latest update record"
      operationId: "getLatestUpdateRecord"
      responses:
        "200":
          description: "Success."
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UpdateRecordResponse"
              examples:
                UpdateRecordResponse:
                  description: "Latest available update"
                  value:
                    updateId: 11256
                    startedTime: "2022-04-05T19:10:17.155325Z"
                    finishedTime: "2022-04-05T19:51:40.327389Z"
                    dataDownloadUrl: "https://download.geofabrik.de/europe/slovakia-latest.osm.pbf"
                    status: "FINISHED"
                    trigger: "SCHEDULED_STARTUP_UPDATE"
                    description: null
                    failedReason: null
                UpdateRecordResponse - not found:
                  description: "No latest update was found"
        "500":
          description: "Internal server error.\n List of possible error codes:<ul><li>LS0004\
            \ - PostgreSQL JDBC connection failure</li></ul>"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                LS0004:
                  description: "LS0004"
                  value:
                    errorMessage:
                      referenceId: "efee2335-7e32-4be8-bd3b-65f2d06b6970"
                      message: "PostgreSQL JDBC connection failure. (cause: Failed\
                        \ to obtain JDBC Connection; nested exception is org.postgresql.util.PSQLException:\
                        \ FATAL: database \"xyz\" does not exist)"
                    businessExceptions:
                    - code: "LS0004"
                      message: "PostgreSQL JDBC connection failure."
                    fieldValidations: []
...
components:
  schemas:
    ManualUpdateRequest:
      type: "object"
      properties:
        url:
          type: "string"
        description:
          type: "string"
    UpdateRecordResponse:
      type: "object"
      properties:
        updateId:
          type: "integer"
          format: "int64"
        startedTime:
          type: "string"
          format: "date-time"
        finishedTime:
          type: "string"
          format: "date-time"
        dataDownloadUrl:
          type: "string"
        status:
          type: "string"
          enum:
          - "RUNNING"
          - "FINISHED"
          - "FAILED"
        trigger:
          type: "string"
          enum:
          - "SCHEDULED_UPDATE"
          - "SCHEDULED_STARTUP_UPDATE"
          - "SCHEDULED_RETRY_UPDATE"
          - "MANUAL_UPDATE"
        description:
          type: "string"
        failedReason:
          type: "string"
    BusinessException:
      type: "object"
      properties:
        code:
          type: "string"
          example: "LSXXXX"
        message:
          type: "string"
          example: "Error message"
    ErrorMessage:
      type: "object"
      properties:
        referenceId:
          type: "string"
          example: "5153c9b9-151d-47e5-93ef-a67ac8663163"
        message:
          type: "string"
          example: "Error message (cause: Cause of this error if available)"
    ErrorResponse:
      required:
      - "businessExceptions"
      - "errorMessage"
      - "fieldValidations"
      type: "object"
      properties:
        errorMessage:
          $ref: "#/components/schemas/ErrorMessage"
        businessExceptions:
          type: "array"
          items:
            $ref: "#/components/schemas/BusinessException"
        fieldValidations:
          type: "array"
          items:
            $ref: "#/components/schemas/FieldValidation"
    FieldValidation:
      type: "object"
      properties:
        field:
          type: "string"
          example: "Field name"
        message:
          type: "string"
          example: "Reason why field validation failed"
    GetUpdateRecordsResponse:
      type: "object"
      properties:
        records:
          type: "array"
          description: "List of all available requested records."
          items:
            $ref: "#/components/schemas/UpdateRecordResponse"
        recordsCount:
          type: "integer"
          description: "Total count of available requested records."
          format: "int64"
          example: 543
...